KFLUXINFRA-3510: move HTTP metrics to middleware package and add unit tests
