+++ internal/http/middleware/middleware_suite_test.go
@@ -1,0 +1,55 @@
+package middleware_test
+
+import (
+	"testing"
+
+	"github.com/prometheus/client_golang/prometheus"
+	dto "github.com/prometheus/client_model/go"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestMiddleware(t *testing.T) {
+	t.Parallel()
+
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Middleware Suite")
+}
+
+func metricFamilyNames(families []*dto.MetricFamily) []string {
+	names := make([]string, len(families))
+	for i, f := range families {
+		names[i] = f.GetName()
+	}
+	return names
+}
+
+func findFamily(reg *prometheus.Registry, name string) *dto.MetricFamily {
+	families, err := reg.Gather()
+	Expect(err).NotTo(HaveOccurred())
+
+	for _, f := range families {
+		if f.GetName() == name {
+			return f
+		}
+	}
+	Fail("metric family not found: " + name)
+	return nil
+}
+
+func labelMap(m *dto.Metric) map[string]string {
+	labels := make(map[string]string)
+	for _, lp := range m.GetLabel() {
+		labels[lp.GetName()] = lp.GetValue()
+	}
+	return labels
+}
+
+func bucketBoundaries(buckets []*dto.Bucket) []float64 {
+	bounds := make([]float64, len(buckets))
+	for i, b := range buckets {
+		bounds[i] = b.GetUpperBound()
+	}
+	return bounds
+}
