+++ metrics_test.go
@@ -2,10 +2,19 @@ package main
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
+	dto "github.com/prometheus/client_model/go"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
+
+func metricFamilyNames(families []*dto.MetricFamily) []string {
+	names := make([]string, len(families))
+	for i, f := range families {
+		names[i] = f.GetName()
+	}
+	return names
+}
 
 var _ = Describe("InitRegistry", func() {
 	It("registers process collector on a fresh registry", func() {
